% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\docType{package}
\name{arcMS-package}
\alias{arcMS}
\alias{arcMS-package}
\title{UNIFI MS converter to Parquet and HDF5 formats}
\description{
\Sexpr[results=text,echo=FALSE]{packageDescription("arcMS", fields = "Description")}
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/leesulab/arcMS/}
  \item \url{https://leesulab.github.io/arcMS/}
}

}
\author{
\strong{Maintainer}: Julien Le Roux \email{julien.le-roux@u-pec.fr} (\href{https://orcid.org/0000-0003-0245-8536}{ORCID})

Authors:
\itemize{
  \item Julien Sade \email{julien.sade@u-pec.fr}
}

}
