% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collect_all_samples_data.R
\name{collect_all_samples_data}
\alias{collect_all_samples_data}
\title{collect_all_samples_data}
\usage{
collect_all_samples_data(connection_params, analysis_id, format = "parquet")
}
\arguments{
\item{connection_params}{Connection parameters to the Unifi API - url and token}

\item{analysis_id}{The id of the analysis}

\item{format}{The format chosen for the exported filed (parquet or HDF5)}
}
\value{
A ataframe for each sample is saved in parquet or HDF5 format in a folder named after the Analysis.
}
\description{
The function collects all samples data from an Analysis using the collect_one_sample_data() function and save them on disk.
}
