% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collect_one_sample_data.R
\name{collect_one_sample_data}
\alias{collect_one_sample_data}
\title{collect_one_sample_data}
\usage{
collect_one_sample_data(
  connection_params,
  sample_id,
  sample_name,
  analysis_name,
  format = "parquet",
  num_spectras = NULL
)
}
\arguments{
\item{connection_params}{Connection parameters to the Unifi API - url and token}

\item{sample_id}{The id of the sample to be collected}

\item{sample_name}{The sample name (to be used for naming the saved file)}

\item{analysis_name}{The name of the Analysis (to be used for naming the folder)}

\item{format}{The format chosen for the exported filed (parquet or HDF5)}

\item{num_spectras}{Number of spectras to be downloaded (OPTIONAL, only if whole sample data not needed)}
}
\value{
A dataframe of the sample's deserialized spectral data is saved in parquet or HDF5 format in the analysis name folder.
}
\description{
The function collects spectral data from a sample in a Unifi Analysis.
First, asynchronous queries are performed for each sample.
Then, the data is deserialized by the function deserialize_data().
The S3 object containing the spectral data is adapted into a dataframe by the function outputlist_to_df().
Finally, the data is saved on disk to the parquet or HDF5 format.
}
