% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_bin_to_ccs.R
\name{convert_bin_to_ccs}
\alias{convert_bin_to_ccs}
\title{Convert bin value to CCS (Collision Cross Section) value
This function converts bin values to CCS values for the given sample.
It is recommended to perform this conversion after peak detection, which can be effectively done using DEIMoS.
For more information on DEIMoS, see \url{https://deimos.readthedocs.io/en/latest/}.}
\usage{
convert_bin_to_ccs(data, connection_params = NULL)
}
\arguments{
\item{data}{An Arrow table object containing the sample data.
This data should include the necessary columns: `bin`, `mz`, and `rt`.
The metadata of this Arrow table should contain the `id` of the sample under `data$metadata$id`.}

\item{connection_params}{OPTIONAL: Connection parameters object created by the
\code{\link{create_connection_params}} function. If not provided, the
\code{\link{get_connection_params}} will look for such an object in the global environment.}
}
\value{
An Arrow table that includes the original data along with an additional column `CCS` containing the CCS values.
  The returned Arrow table will also retain the original metadata.
}
\description{
Convert bin value to CCS (Collision Cross Section) value
This function converts bin values to CCS values for the given sample.
It is recommended to perform this conversion after peak detection, which can be effectively done using DEIMoS.
For more information on DEIMoS, see \url{https://deimos.readthedocs.io/en/latest/}.
}
