% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_drifttime_bin.R
\name{convert_bin_to_drifttime}
\alias{convert_bin_to_drifttime}
\title{Convert Bin Numbers to Drift Time}
\usage{
convert_bin_to_drifttime(sample_id, bins, connection_params = NULL)
}
\arguments{
\item{sample_id}{The identifier of the sample result.}

\item{bins}{An integer vector of bin numbers to convert.}

\item{connection_params}{OPTIONAL: Connection parameters object created by the
\code{\link{create_connection_params}} function.}
}
\value{
A numeric vector of drift time values (in milliseconds) corresponding to the input bins.
}
\description{
This function converts bin numbers to drift time values by querying the
UNIFI API endpoint \code{convertbintodrifttime}.
}
