% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_drifttime_bin.R
\name{convert_drifttime_to_bin}
\alias{convert_drifttime_to_bin}
\title{Convert Drift Time to Bin Number}
\usage{
convert_drifttime_to_bin(sample_id, drift_times, connection_params = NULL)
}
\arguments{
\item{sample_id}{The identifier of the sample result.}

\item{drift_times}{A numeric vector of drift time values (in milliseconds) to convert.}

\item{connection_params}{OPTIONAL: Connection parameters object created by the
\code{\link{create_connection_params}} function.}
}
\value{
An integer vector of bin numbers corresponding to the input drift times.
}
\description{
This function converts drift time values to bin numbers by querying the
UNIFI API endpoint \code{convertdrifttimetobin}.
}
