% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_one_sample_data.R
\name{convert_one_sample_data}
\alias{convert_one_sample_data}
\title{Convert one sample}
\usage{
convert_one_sample_data(
  connection_params,
  sample_id,
  sample_name,
  analysis_name,
  format = "parquet",
  num_spectras = NULL
)
}
\arguments{
\item{connection_params}{Connection parameters to the Unifi API - url and token}

\item{sample_id}{The id of the sample to be collected}

\item{sample_name}{The sample name (to be used for naming the saved file)}

\item{analysis_name}{The name of the Analysis (to be used for naming the folder)}

\item{format}{The format chosen for the exported file (parquet or hdf5)}

\item{num_spectras}{Number of spectras to be downloaded (OPTIONAL, only if whole sample data not needed)}
}
\value{
Datatables of the sample's spectral data and metadata are saved in Parquet or HDF5 format in the analysis name folder.
}
\description{
The function collects spectral data from a sample in a Unifi Analysis.
First, asynchronous queries are performed for each sample.
Then, the data is deserialized by the \code{\link{deserialize_data}} function.
The S3 object containing the spectral data is adapted into a dataframe by the \code{\link{outputlist_to_df}} function.
Finally, the data is saved on disk to the Parquet or HDF5 format with the \code{\link{save_one_sample_data}} function.
}
\seealso{
\code{\link{collect_one_sample_data}} for only collecting data by dowloading from the API into the R environment, and \code{\link{save_one_sample_data}} to save collected data from the R environment to Parquet or HDF5 files.
}
