% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_connection.R
\name{create_connection_params}
\alias{create_connection_params}
\title{Create connection parameters}
\usage{
create_connection_params(
  identityurl = "http://localhost:50333/identity/connect/token",
  username = "administrator",
  password = "administrator",
  apihosturl = "http://localhost:50034/unifi/v1",
  install = TRUE,
  verbose = TRUE
)
}
\arguments{
\item{identityurl}{The \code{url} to connect to the UNIFI API identity server.}

\item{username}{The \code{username} to connect to the UNIFI API identity server.}

\item{password}{The \code{password} to connect to the UNIFI API identity server.}

\item{apihosturl}{The \code{url} to connect to the UNIFI API server (host).}

\item{install}{if TRUE, will install the token in your \code{.Renviron} file for use in future sessions.  Defaults to TRUE.}

\item{verbose}{if TRUE, will print messages.  Defaults to TRUE.}
}
\value{
A list containing all parameters needed for the connection, in a \code{\link{connection_params}} object.
}
\description{
Create connection parameters
}
\details{
\code{create_connection_params} is an utility function that returns the url and token for UNIFI API connection
}
