% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_dataset.R
\name{create_sample_dataset}
\alias{create_sample_dataset}
\title{Create a sample dataset from a Parquet file}
\usage{
create_sample_dataset(file, method = "inram")
}
\arguments{
\item{file}{A character file name or URI of Parquet file.}

\item{method}{Whether to import the data in RAM or keep it on-disk}
}
\value{
A \code{\link{sample_dataset}} object, containing the sample data,
sample metadata and spectrum metadata datatables.
}
\description{
The function creates a sample dataset object from data imported from a Parquet file.
'
}
\seealso{
\code{\link{save_one_sample_data}} to save collected data from the R environment to Parquet or HDF5 files, and \code{\link{convert_one_sample_data}} to both collect data and saving to files.
}
