% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/decodeVarint.R
\name{DecodeVarint32}
\alias{DecodeVarint32}
\title{DecodeVarint32}
\usage{
DecodeVarint32(buffer, pos)
}
\arguments{
\item{buffer}{The buffer data}

\item{pos}{The position of the beginning of the message}
}
\value{
Decoded message position
}
\description{
A function to decode a buffer data.
}
\references{
\itemize{ Translated from Python:
  \item Encoder - https://github.com/protocolbuffers/protobuf/blob/master/python/google/protobuf/internal/encoder.py
  \item Decoder - https://raw.githubusercontent.com/protocolbuffers/protobuf/master/python/google/protobuf/internal/decoder.py
  \item Author - Fred K. Gruber <fred@gnshealthcare.com>
}
}
\keyword{internal}
