% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deserialize_data.R
\name{deserialize_data}
\alias{deserialize_data}
\title{Deserialize data}
\usage{
deserialize_data(rg_content)
}
\arguments{
\item{rg_content}{the $content of a request in octet-stream}
}
\value{
A S3 object which contains the spectral data associated with each spectrum
}
\description{
A function to deserialize the protobuf data
}
\keyword{internal}
