% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/explode_spectra.R
\name{explode_spectra}
\alias{explode_spectra}
\title{Explode Spectra Data}
\usage{
explode_spectra(wide_df)
}
\arguments{
\item{wide_df}{A wide format dataframe with columns 'masses', 'intensities', and 'scan_size'.}
}
\value{
A long format dataframe with the exploded spectral data.
}
\description{
This function takes a wide format dataframe and unnests (explodes) it into a long format.
The unnested table contains the columns rt (retention time), scanid (id of each scan),
energy_level (low or high), mz (m/z values) and intensities.
}
