% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_spectra.R
\name{get_chromatogram_data}
\alias{get_chromatogram_data}
\title{Retrieve Chromatogram Data from a Sample Result}
\usage{
get_chromatogram_data(sample_id, chromatogram_id, connection_params = NULL)
}
\arguments{
\item{sample_id}{The identifier of the sample result.}

\item{chromatogram_id}{The identifier of the chromatogram.}

\item{connection_params}{OPTIONAL: Connection parameters object.}
}
\value{
A data.table with columns \code{retentionTimes} and \code{intensities}.
}
\description{
Retrieve Chromatogram Data from a Sample Result
}
