% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_spectra.R
\name{get_chromatogram_infos}
\alias{get_chromatogram_infos}
\title{Retrieve Chromatogram Information from a Sample Result}
\usage{
get_chromatogram_infos(sample_id, connection_params = NULL)
}
\arguments{
\item{sample_id}{The identifier of the sample result.}

\item{connection_params}{OPTIONAL: Connection parameters object.}
}
\value{
A data.table with chromatogram metadata (id, name, type, etc.).
}
\description{
Retrieve Chromatogram Information from a Sample Result
}
