% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_components.R
\name{get_components}
\alias{get_components}
\title{Retrieve Components from a Sample Result in UNIFI API}
\usage{
get_components(sample_id, connection_params = NULL)
}
\arguments{
\item{sample_id}{The identifier of the sample result for which components are to be retrieved.}

\item{connection_params}{OPTIONAL: Connection parameters object created by the
\code{\link{create_connection_params}} function. If not provided, the
\code{\link{get_connection_params}} will look for such object in the global environment.}
}
\value{
A data.table containing component information from the UNIFI API,
  including columns such as name, id, componentStatus, chromatographic.retentionTime,
  ms.massToChargeRatio, ims.driftTime, ims.collisionCrossSection, etc.
}
\description{
This function retrieves component information from the UNIFI API for a specified
sample result. Components include identified and candidate markers with their
associated chromatographic, mass spectrometric, and ion mobility data.
}
