% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_connection.R
\name{get_connection_params}
\alias{get_connection_params}
\title{Retrieve the connection parameters object}
\usage{
get_connection_params(envir = parent.frame())
}
\arguments{
\item{envir}{The environment to look for object containing connection parameters}
}
\description{
This function will retrieve the connection parameters object from
the global environment, created by the \code{\link{create_connection_params}} function
}
