% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collect_all_samples_data.R
\name{get_sample_list}
\alias{get_sample_list}
\title{get_sample_list}
\usage{
get_sample_list(connection_params, analysis_id)
}
\arguments{
\item{connection_params}{- Connection parameters to the Unifi API - url and token}

\item{analysis_id}{- The id of the analysis}
}
\value{
A dataframe wih sample names and IDs, along with the analysis name.
}
\description{
The function connects to a Unifi analysis and gets the list of all samples. It returns a dataframe with sample names and IDs, along with the analysis name.
}
