% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_samples_list.R
\name{get_samples_list}
\alias{get_samples_list}
\title{Retrieve the list of Samples Results from an Analysis in UNIFI API}
\usage{
get_samples_list(analysis_id, connection_params = NULL)
}
\arguments{
\item{analysis_id}{The identifier of the Analysis}

\item{connection_params}{OPTIONAL: Connection parameters object created by the
\code{\link{create_connection_params}} function. If not provided, the
\code{\link{get_connection_params}} will look for such object in the global environment}
}
\value{
A dataframe wih sample names and their IDs, along with the analysis name and sample metadata.
}
\description{
The function connects to a UNIFI Analysis and gets the list of all samples.
It returns a dataframe with sample names and IDs, along with the analysis name.
}
