% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_spectra.R
\name{get_spectra_by_rt}
\alias{get_spectra_by_rt}
\title{Retrieve Individual MS Spectra from UNIFI API}
\usage{
get_spectra_by_rt(
  sample_id,
  rt_min,
  rt_max,
  energy_level = "Low",
  connection_params = NULL
)
}
\arguments{
\item{sample_id}{The identifier of the sample result.}

\item{rt_min}{Minimum retention time (in minutes) for the filter.}

\item{rt_max}{Maximum retention time (in minutes) for the filter.}

\item{energy_level}{Energy level to filter: \code{"Low"} for MS1, \code{"High"} for MS2.}

\item{connection_params}{OPTIONAL: Connection parameters object created by the
\code{\link{create_connection_params}} function.}
}
\value{
A list of deserialized spectrum objects, each containing masses, intensities,
  scan sizes, retention time, energy level, and ionization polarity.
}
\description{
This function retrieves individual mass spectra from the UNIFI API for a given
sample, filtered by retention time range and energy level (Low for MS1,
High for MS2/MSe). The spectra are returned as protobuf binary data and
deserialized into R data structures.
}
\details{
This is the R equivalent of the JavaScript \code{loadSpectra} function from the
UNIFI explorer app, which queries the
\code{sampleresults({id})/spectra/mass.mse} endpoint.
}
