% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_spectra.R
\name{get_spectra_count}
\alias{get_spectra_count}
\title{Count Available Spectra for a Sample}
\usage{
get_spectra_count(sample_id, connection_params = NULL)
}
\arguments{
\item{sample_id}{The identifier of the sample result.}

\item{connection_params}{OPTIONAL: Connection parameters object.}
}
\value{
The total number of spectra available for the sample.
}
\description{
Count Available Spectra for a Sample
}
