% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_spectrum_infos.R
\name{get_spectrum_infos}
\alias{get_spectrum_infos}
\title{Retrieve Spectrum Information from Unifi API}
\usage{
get_spectrum_infos(connection_params, sample_id)
}
\arguments{
\item{connection_params}{Connection parameters to the Unifi API, including the API host URL and access token.}

\item{sample_id}{The identifier of the sample result for which spectrum information is to be retrieved.}
}
\value{
A list containing two elements: a dataframe with detailed spectrum information from the Unifi API,
        and a formatted JSON string of the spectrum information.

Example usage:
con_params <- create_connection_params(api_host_url, api_token)
results <- get_spectrum_infos(con_params, sample_result_id)
spectrum_infos_df <- results$spectrum_infos
formatted_json <- results$formatted_json
}
\description{
This function retrieves spectrum information from the Unifi API for a specified sample result using the provided connection parameters.
It extracts detailed spectrum information associated with the given sample result identifier.
The function returns both a dataframe containing the spectrum information and a formatted JSON string of the same data.
}
