% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_connection.R
\name{get_unifi_api_token}
\alias{get_unifi_api_token}
\title{Retrieve the API token in the environment variable}
\usage{
get_unifi_api_token()
}
\description{
This function will retrieve the UNIFI API token from the OS environment
}
