% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mzml_to_parquet.R
\name{mzml_to_dt}
\alias{mzml_to_dt}
\title{Convert mzML (or mzXML) file to data.table in long format}
\usage{
mzml_to_dt(filename, chunksize = 10000)
}
\arguments{
\item{filename}{Path name of the netCDF, mzXML or mzML file to read/write.}

\item{chunksize}{size (number of scans) for each chunk to be retrieved from the mzML file.}
}
\value{
A \code{\link{data.table}} object containing the spectra
}
\description{
The function collects spectral data and metadata from an mzML file with mzR package.
Data are transformed to data.table in long format.
}
\seealso{
\code{\link{mzml_to_sample_dataset}} to convert mzML data to sample_dataset object.
}
