% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mzml_to_parquet.R
\name{mzml_to_sample_dataset}
\alias{mzml_to_sample_dataset}
\title{Convert mzML (or mzXML) file to sample_dataset}
\usage{
mzml_to_sample_dataset(filename, chunksize = 10000)
}
\arguments{
\item{filename}{Path name of the netCDF, mzXML or mzML file to read/write.}

\item{chunksize}{size (number of scans) for each chunk to be retrieved from the mzML file.}
}
\value{
A \code{\link{sample_dataset}} object containing the sample spectra
and sample metadata datatables.
}
\description{
The function collects spectral data and metadata from an mzML file with mzR package.
Data are transformed to data.table and saved in a sample_dataset object.
Metadata are arranged as a data.frame and passed to sample_metadata slot.
}
\seealso{
\code{\link{convert_mzml_to_parquet}} to directly save converted data to Parquet file.
}
