% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outputlist_to_df.R
\name{outputlist_to_df}
\alias{outputlist_to_df}
\title{outputlist_to_df}
\usage{
outputlist_to_df(outputlist)
}
\arguments{
\item{outputlist}{Deserialized data in S3 format obtained with the deserialized_data() function}
}
\value{
A dataframe containing the spectral data associated with each spectra.
}
\description{
The function adapts the data in S3 format obtained with the `deserialize_data()` function into a dataframe.
}
