% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\docType{package}
\name{parquetMS-package}
\alias{parquetMS}
\alias{parquetMS-package}
\title{Unifi MS converter to Parquet and HDF5 formats.}
\description{
\Sexpr[results=text,echo=FALSE]{packageDescription("parquetMS", fields = "Description")}
}
\author{
\strong{Maintainer}: Julien Le Roux \email{julien.le-roux@u-pec.fr}

Authors:
\itemize{
  \item Julien Sade \email{julien.sade@u-pec.fr}
}

}
