% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_app.R
\name{run_app}
\alias{run_app}
\title{Run Shiny application to use the conversion package}
\usage{
run_app()
}
\description{
The function launches the Shiny app. The app can be used to connect to the UNIFI API,
navigate in folders and analyses available in UNIFI, select samples or analysis to convert,
and run the conversion. The destination folder of converted files
and their format (Parquet or HDF5) can be chosen.
}
