% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_dataset.R
\docType{class}
\name{sample_dataset-class}
\alias{sample_dataset-class}
\alias{sample_dataset}
\alias{get_sample_data,sample_dataset-method}
\alias{get_sample_data}
\title{Class containing a sample data and metadata}
\usage{
\S4method{get_sample_data}{sample_dataset}(obj)
}
\arguments{
\item{obj}{The \code{\link{sample_dataset}} object to access.}
}
\value{
\code{get_sample_data} returns a data.table object containing the sample data.
}
\description{
Contains sample data (as data.table in RAM or as pointer to Parquet file),
metadata and spectrum metadata in table and json formats.
}
\details{
Objects for this class are returned by \code{\link{collect_one_sample_data}}
and \code{\link{create_sample_dataset}}.
}
\section{Methods (by generic)}{
\itemize{
\item \code{get_sample_data(sample_dataset)}: Accessor method to obtain the sample_data table.

}}
\section{Slots}{

\describe{
\item{\code{sample_data}}{Contains a \code{datatable} with the sample data, or
an \code{Arrow Dataset} R6 object pointing to the Parquet file data.}

\item{\code{sample_metadata}}{Contains a \code{datatable} with the sample metadata.}

\item{\code{spectrum_metadata}}{Contains a \code{datatable} with the spectrum metadata.}

\item{\code{sample_metadata_json}}{Contains a \code{character} with the sample metadata.}

\item{\code{spectrum_metadata_json}}{Contains a \code{character} with the spectrum metadata.}
}}

\section{Use the \code{\link{collect_one_sample_data}} to}{

  store the sample data, metadata and spectrum metadata, in table and json formats.
}

