% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_dataset.R
\docType{class}
\name{sample_dataset_ondisk-class}
\alias{sample_dataset_ondisk-class}
\alias{sample_dataset_ondisk}
\title{Class containing a sample data as pointer to Parquet file, and metadata}
\arguments{
\item{obj}{The \code{\link{sample_dataset}} object to access.}
}
\description{
Contains sample data as a pointer to a Parquet file, and sample metadata
and spectrum metadata in table and json formats.
}
\details{
Objects for this class can be created by \code{\link{create_sample_dataset}}
with the ondisk method.
}
\section{Slots}{

\describe{
\item{\code{sample_data}}{Contains an \code{Arrow Dataset} R6 object pointing to the Parquet file.}

\item{\code{sample_metadata}}{Contains a \code{datatable} with the sample metadata.}

\item{\code{spectrum_metadata}}{Contains a \code{datatable} with the spectrum metadata.}

\item{\code{sample_metadata_json}}{Contains a \code{character} with the sample metadata.}

\item{\code{spectrum_metadata_json}}{Contains a \code{character} with the spectrum metadata.}
}}

\section{Use the \code{\link{collect_one_sample_data}} to}{

  store the sample data, metadata and spectrum metadata, in table and json formats.
}

