% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_infos.R
\docType{class}
\name{sample_infos-class}
\alias{sample_infos-class}
\alias{sample_infos}
\alias{get_sample_metadata,sample_infos-method}
\alias{get_sample_metadata}
\alias{get_sample_name,sample_infos-method}
\alias{get_sample_name}
\alias{get_analysis_name,sample_infos-method}
\alias{get_analysis_name}
\alias{get_sample_metadata_json,sample_infos-method}
\alias{get_sample_metadata_json}
\title{Class containing a sample's information}
\usage{
\S4method{get_sample_metadata}{sample_infos}(obj)

\S4method{get_sample_name}{sample_infos}(obj)

\S4method{get_analysis_name}{sample_infos}(obj)

\S4method{get_sample_metadata_json}{sample_infos}(obj)
}
\arguments{
\item{obj}{The \code{\link{sample_infos}} object to access.}
}
\value{
\code{get_sample_metadata} returns a data.frame object containing the sample metadata.

\code{get_sample_metadata} returns a character object containing the sample name.

\code{get_analysis_name} returns a character object containing the analysis name.

\code{get_sample_metadata_json} returns a json character object containing the sample metadata.
}
\description{
Contains sample metadata in table and json formats.
}
\details{
Objects for this class are returned by \code{\link{get_sample_infos}}.
}
\section{Methods (by generic)}{
\itemize{
\item \code{get_sample_metadata(sample_infos)}: accessor method to obtain the sample_metadata table.

\item \code{get_sample_name(sample_infos)}: accessor method to obtain the sample name.

\item \code{get_analysis_name(sample_infos)}: accessor method to obtain the analysis name.

\item \code{get_sample_metadata_json(sample_infos)}: accessor method to obtain the sample metadata in json format.

}}
\section{Slots}{

\describe{
\item{\code{sample_metadata}}{Contains a \code{dataframe} with the sample metadata.}

\item{\code{formatted_json}}{Contains a \code{character} with the sample metadata.}
}}

\section{Use the \code{\link{get_sample_infos}} to}{

  store the sample metadata, in table and json formats.
}

