% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_search.R
\name{sample_search}
\alias{sample_search}
\title{Retrieve Sample Results from Unifi API}
\usage{
sample_search(connection_params, analysis_id)
}
\arguments{
\item{connection_params}{Connection parameters to the Unifi API, including the API host URL and access token.}

\item{analysis_id}{The identifier of the analysis for which sample results are to be retrieved.}
}
\value{
A data frame containing samples result information from the Unifi API,
        with superfluous columns omitted for clarity.

Example usage:
con_params <- create_connection_params(api_host_url, api_token)
sample_results <- sample_search(con_params, analysis_id)
}
\description{
This function retrieves samples result information from the Unifi API for a specified analysis using the provided connection parameters.
}
