% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_one_sample_data.R
\name{save_one_sample_data}
\alias{save_one_sample_data}
\title{Save collected data from one sample}
\usage{
save_one_sample_data(
  sample_dataset,
  sample_name = NULL,
  analysis_name = NULL,
  format = "parquet"
)
}
\arguments{
\item{sample_dataset}{Collected data from a sample, stored in a \code{\link{sample_dataset}} object}

\item{sample_name}{The sample name (to be used for naming the saved file)}

\item{analysis_name}{The name of the Analysis (to be used for naming the folder)}

\item{format}{The format chosen for the exported file (parquet or hdf5)}
}
\value{
Datatables of the sample's spectral data and metadata are saved in Parquet or HDF5 format in the analysis name folder.
}
\description{
The function saves data collected from one sample with the \code{\link{collect_one_sample_data}} function.
Datatables of spectra and metadata are saved in either two Parquet files
or one HDF5 file.
}
\seealso{
\code{\link{collect_one_sample_data}} for only collecting data by dowloading from the API into the R environment, and \code{\link{convert_one_sample_data}} to both collect data and saving to files.
}
