% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_connection.R
\name{send_request}
\alias{send_request}
\title{Check if the connection is still active before current request then return request}
\usage{
send_request(request, connection_params)
}
\arguments{
\item{request}{The request to be executed during/after checking}

\item{connection_params}{The current connection parameters object (with current token)}
}
\description{
This function will check if the connection is still active
and the current token is valid before retrieving request results
}
