% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis_search.R
\name{analysis_search}
\alias{analysis_search}
\title{Retrieve Analyses from Unifi API}
\usage{
analysis_search(connection_params, folder_id)
}
\arguments{
\item{connection_params}{Connection parameters to the Unifi API, including the API host URL and access token.}

\item{folder_id}{The identifier of the folder for which items are to be retrieved.}
}
\value{
A data frame containing analysis item information from the Unifi API,
        with unnecessary columns removed for clarity.

Example usage:
con_params <- create_connection_params(api_host_url, api_token)
analysis_items <- analysis_search(con_params, folder_id)
}
\description{
This function retrieves items information from the Unifi API for a specified Unifi folder using the provided connection parameters.
The function filters and returns analysis items based on their type.
}
