% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_one_sample_data.R
\name{collect_one_sample_data}
\alias{collect_one_sample_data}
\title{Collect data from one sample}
\usage{
collect_one_sample_data(connection_params, sample_id, num_spectras = NULL)
}
\arguments{
\item{connection_params}{Connection parameters to the Unifi API - url and token}

\item{sample_id}{The id of the sample to be collected}

\item{num_spectras}{Number of spectras to be downloaded (OPTIONAL, only if whole sample data not needed)}
}
\value{
A \code{\link{sample_dataset}} object, containing the sample data,
sample metadata and spectrum metadata datatables.
}
\description{
The function collects spectral data from a sample in a Unifi Analysis.
First, asynchronous queries are performed for each sample.
Then, the data is deserialized by the function deserialize_data().
The S3 object containing the spectral data is adapted into a dataframe by the function outputlist_to_df().
}
\seealso{
\code{\link{save_one_sample_data}} to save collected data from the R environment to Parquet or HDF5 files, and \code{\link{convert_one_sample_data}} to both collect data and saving to files.
}
