% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_connection.R
\docType{class}
\name{connection_params-class}
\alias{connection_params-class}
\alias{connection_params}
\alias{connection_apihosturl,connection_params-method}
\alias{connection_apihosturl}
\alias{connection_token,connection_params-method}
\alias{connection_token}
\title{Class containing UNIFI API connection parameters}
\usage{
\S4method{connection_apihosturl}{connection_params}(obj)

\S4method{connection_token}{connection_params}(obj)
}
\arguments{
\item{obj}{The \code{\link{connection_params}} object to access.}
}
\value{
\code{connection_apihosturl} returns a character object containing the connection url.

\code{connection_token} returns a character object containing the connection token.
}
\description{
Contains url and token for connection to UNIFI API.
}
\details{
Objects for this class are returned by \code{\link{create_connection_params}}.
}
\section{Methods (by generic)}{
\itemize{
\item \code{connection_apihosturl(connection_params)}: Accessor method to obtain the connection url.

\item \code{connection_token(connection_params)}: Accessor method to obtain the connection token.

}}
\section{Slots}{

\describe{
\item{\code{identityurl}}{Contains a \code{character} with the URL for connection to the UNIFI identity server.}

\item{\code{username}}{Contains a \code{character} with the user name for connection to the UNIFI identity server (e.g. administrator).}

\item{\code{password}}{Contains a \code{character} with the user password for connection to the UNIFI identity server (e.g. administrator).}

\item{\code{apihosturl}}{Contains a \code{character} with the URL for connection to the UNIFI API server (host).}

\item{\code{token}}{A \code{character} with the token sent back from the UNIFI API.}
}}

\section{Use the \code{\link{create_connection_params}} to}{

  store the connection parameters,
  using the default parameters or replacing them with the needed URLs,
  username and password to the UNIFI API. It will request a connection
  token that will be used by other functions to retrieve data.
}

