% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CCS_collect.R
\name{convert_bin_to_ccs}
\alias{convert_bin_to_ccs}
\title{Convert bin data to CCS (Collision Cross Sections)}
\usage{
convert_bin_to_ccs(sample_id, unnestdt, connection_params = NULL)
}
\arguments{
\item{sample_id}{The id of the sample to be collected.}

\item{unnestdt}{A data frame containing the sample data, expected to include columns `bin`, `mz`, and `rt`.}

\item{connection_params}{OPTIONAL: Connection parameters object created by the
\code{\link{create_connection_params}} function. If not provided, the
\code{\link{get_connection_params}} will look for such an object in the global environment.}
}
\value{
A data frame with an additional column `CCS` containing the converted values.
}
\description{
Convert bin data to CCS (Collision Cross Sections)
}
