% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_one_sample_data.R
\name{convert_one_sample_data}
\alias{convert_one_sample_data}
\title{Convert one sample}
\usage{
convert_one_sample_data(
  sample_id,
  connection_params = NULL,
  format = "parquet",
  path = NULL,
  overwrite = T,
  num_spectras = NULL
)
}
\arguments{
\item{sample_id}{The id of the sample to be collected}

\item{connection_params}{OPTIONAL: Connection parameters object created by the
\code{\link{create_connection_params}} function. If not provided, the
\code{\link{get_connection_params}} will look for such object in the global environment}

\item{format}{The format chosen for the exported file (Parquet or HDF5)}

\item{path}{OPTIONAL The destination path for the exported file}

\item{num_spectras}{OPTIONAL Number of spectras to be downloaded (OPTIONAL, only if whole sample data not needed, e.g. for testing purposes)}
}
\value{
Datatables of the sample's spectral data and metadata are saved in Parquet or HDF5 format in the analysis name folder.
}
\description{
The function collects spectral data from a sample in a UNIFI Analysis.
First, asynchronous queries are performed for each sample.
Then, the data is deserialized by the \code{\link{deserialize_data}} function.
The S3 object containing the spectral data is adapted into a dataframe by the \code{\link{outputlist_to_df}} function.
Finally, the data is saved on disk to the Parquet or HDF5 format with the \code{\link{save_one_sample_data}} function.
}
\seealso{
\code{\link{collect_one_sample_data}} for only collecting data by dowloading from the API into the R environment, and \code{\link{save_one_sample_data}} to save collected data from the R environment to Parquet or HDF5 files.
}
