% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/folders_search.R
\name{folders_search}
\alias{folders_search}
\title{Retrieve Folders from Unifi API}
\usage{
folders_search(connection_params = NULL)
}
\arguments{
\item{connection_params}{Connection parameters to the Unifi API, including the API host URL and access token.}
}
\value{
A data frame containing folder information from the Unifi API.
}
\description{
This function retrieves folder information from the Unifi API using the provided connection parameters.
}
