% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_infos.R
\name{get_sample_infos}
\alias{get_sample_infos}
\title{Retrieve Sample Information from Unifi API with a sample id}
\usage{
get_sample_infos(connection_params, sample_id)
}
\arguments{
\item{connection_params}{Connection parameters to the Unifi API, including the API host URL and access token.}

\item{sample_id}{The identifier of the sample result for which spectrum information is to be retrieved.}
}
\value{
A \code{\link{sample_infos}} object, consisting of a list containing two elements: a dataframe with detailed sample information from the Unifi API, and a formatted JSON string of the sample information.

Example usage:
con_params <- create_connection_params(api_host_url, api_token)
results <- get_sample_infos(con_params, sample_result_id)
}
\description{
This function retrieves spectrum information from the Unifi API for a specified sample result using the provided connection parameters.
It extracts detailed spectrum information associated with the given sample result identifier.
The function returns both a dataframe containing the sample information and a formatted JSON string of the same data.
}
