% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_samples_list.R
\name{get_samples_list}
\alias{get_samples_list}
\title{Retrieve the list of Samples Results from an Analysis in Unifi API}
\usage{
get_samples_list(connection_params, analysis_id)
}
\arguments{
\item{connection_params}{- Connection parameters to the Unifi API - url and token}

\item{analysis_id}{- The identifier of the Analysis}
}
\value{
A dataframe wih sample names and their IDs, along with the analysis name and sample metadata.
}
\description{
The function connects to a Unifi Analysis and gets the list of all samples.
It returns a dataframe with sample names and IDs, along with the analysis name.
}
